Release impersonation work (#16)
* Cherrypick druid segmentation PR https://github.com/apache/incubator-airflow/pull/1771

* [AIRFLOW-413] Fix unset path bug when backfilling via pickle

Please accept this PR that addresses the following issues:
- https://issues.apache.org/jira/browse/AIRFLOW-413

This fixes a bug when a pickled DAG is used in a backfill.

Testing Done:
- Existing unit tests + running a backfill command that used to fail
  before on this error

mistercrunch artwr plypaul

Closes #1723 from aoen/ddavydov/fix_undefined_path_for_backfilling

* [AIRFLOW-341][operators] Add resource requirement attributes to operators

This PR adds optional resource requirements for tasks for use with
resource managers such as Yarn and Mesos.

Considerations:
- I chose to force users to encapsulate resources in a resources object
e.g. Resources(cpu=1) instead of just cpu=1 in their dag attributes.
This creates the pain of having to import Resources for almost every
DAG. I think this is kind of important for scoping/namespacing which we
should start doing.
- Once resources are used by executors we need to add documentation for
these new resources (and examples)

Testing Done:
- New/existing unit tests

plypaul artwr mistercrunch jlowin bolkedebruin criccomini

Closes #1669 from aoen/ddavydov/ddavydov/augment_tasks_with_resources

* [AIRFLOW-398] Use cgroups to limit task resource utilization

* [AIRFLOW-219] Unix impersonations for subtasks

Adds a `run_as_user` parameter to subtasks such
that the `airflow run --raw` command will be
ran with `sudo -u user`. Logging from subtasks
will go to STDOUT instead to be read and logged
by the parent process. Also introduces a default
user to impersonate if `run_as_user` is not
specified.

* [Remove after migration] Ensure log file is writable by airflow_sudoable